requirements-dev.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
databases==0.9.0
exceptiongroup==1.2.2
fastapi==0.114.0
greenlet==3.0.3
h11==0.14.0
idna==3.8
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic-settings==2.4.0
pydantic_core==2.23.3
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
